+++ mvport/__init__.py
@@ -5,3 +5,6 @@
 __author__ = """Rafael Lopes Conde dos Reis"""
 __email__ = 'rafael.lcreis@gmail.com'
 __version__ = '1.0.0'
+
+from stock import Stock
+from portfolio import Portfolio